Separated to different files & combine to one main()
Separated the large files into multiple smaller files for future readability. The whole project now lives under one main function as well. The functionality is chosen using arguments.
