--- client/config.json
@@ -1,1 +1,0 @@
-{"plex_token": "", "plex_server_url": "","server_url":"","username":""}
