+++ config.json
@@ -1,0 +1,8 @@
+{
+    "plex_token": "",
+    "plex_server_url": "",
+    "server_url": "",
+    "username": "",
+    "port": "8080",
+    "interval_minutes":"120"
+}
