--- server/config.json
@@ -1,1 +1,0 @@
-{"port":"8080"}
