+++ user.go
@@ -1,0 +1,7 @@
+package main
+
+type User struct {
+	Firstname string `json:"firstname"`
+	Lastname  string `json:"lastname"`
+	Username  string `json:"username"`
+}
